Pull osg-test and osg-ca-generator from GitHub instead of SVN
(We can't use the AFS repos because they are not accessible from
machines that do not have AFS mounted).

The patch uses `git archive` to create the tarball that gets uploaded to
the test VMs. git archive embeds the commit ID into the tarball
metadata. run-job uses `git get-tar-commit-id` to get the ID out of the
tarball so that the commit ID can be put into the logs. If it's not
already present on the machine, then `run-job` will install git.

git-svn-id: https://vdt.cs.wisc.edu/svn/new-test/trunk/vm-test-runs@22710 4e558342-562e-0410-864c-e07659590f8c