allow empty params files
the intention of the surrounding code is to ignore files that do not
parse as yaml.  However, an empty file under parameters.d gets parsed
as yaml.load("") -> None, and that None is not iterable by set()

So, change this 'if' condition not to raise an exception if
param_contents for a given yaml file is empty/None.